refact: domain.dto refactoring
