+++ recommend/src/main/java/com/lxp/recommend/application/mapper/LearnerProfileMapper.java
@@ -1,8 +1,8 @@
 package com.lxp.recommend.application.mapper;
 
 import com.lxp.recommend.application.port.required.dto.LearnerProfileData;
-import com.lxp.recommend.domain.dto.LearnerLevel;
-import com.lxp.recommend.domain.dto.LearnerProfileView;
+import com.lxp.recommend.domain.model.ids.LearnerLevel;
+import com.lxp.recommend.infrastructure.external.user.LearnerProfileView;
 
 public class LearnerProfileMapper {
