+++ recommend/src/main/java/com/lxp/recommend/domain/model/ids/EnrollmentStatus.java
@@ -1,4 +1,4 @@
-package com.lxp.recommend.domain.dto;
+package com.lxp.recommend.domain.model.ids;
 
 public enum EnrollmentStatus {
     ENROLLED, COMPLETED, CANCELLED
