+++ recommend/src/main/java/com/lxp/recommend/domain/model/ids/LearnerLevel.java
@@ -1,4 +1,4 @@
-package com.lxp.recommend.domain.dto;
+package com.lxp.recommend.domain.model.ids;
 
 import com.lxp.common.enums.Level;
