+++ recommend/src/main/java/com/lxp/recommend/infrastructure/external/common/LevelMapper.java
@@ -1,4 +1,4 @@
-package com.lxp.recommend.domain.dto;
+package com.lxp.recommend.infrastructure.external.common;
 
 import com.lxp.common.enums.Level;
